Spread type definitions across component directories
