+++ packages/fuselage/src/components/Accordion/index.d.ts
@@ -1,0 +1,13 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type AccordionProps = ComponentProps<typeof Box>;
+type AccordionItemProps = Omit<ComponentProps<typeof Box>, 'title'> & {
+  defaultExpanded?: boolean;
+  title?: string;
+};
+
+export const Accordion: ForwardRefExoticComponent<AccordionProps> & {
+  Item: ForwardRefExoticComponent<AccordionItemProps>;
+};
