+++ packages/fuselage/src/components/Badge/index.d.ts
@@ -1,0 +1,11 @@
+import { ElementType, ForwardRefExoticComponent } from 'react';
+
+type BadgeProps = {
+  is?: ElementType;
+  variant?: 'primary' | 'danger' | 'warning';
+  disabled?: boolean;
+  className?: string;
+  children?: any;
+  title?: any;
+};
+export const Badge: ForwardRefExoticComponent<BadgeProps>;
