+++ packages/fuselage/src/components/Box/Scrollable/index.d.ts
@@ -1,0 +1,8 @@
+import { ForwardRefExoticComponent, PropsWithChildren } from 'react';
+
+type ScrollableProps = PropsWithChildren<{
+  horizontal?: boolean;
+  vertical?: boolean;
+  onScrollContent?: (touching: { top: boolean }) => void;
+}>;
+export const Scrollable: ForwardRefExoticComponent<ScrollableProps>;
