+++ packages/fuselage/src/components/Callout/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type CalloutProps = ComponentProps<typeof Box>;
+export const Callout: ForwardRefExoticComponent<CalloutProps>;
