+++ packages/fuselage/src/components/CheckBox/index.d.ts
@@ -1,0 +1,8 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type CheckBoxProps = ComponentProps<typeof Box> & {
+  indeterminate?: boolean;
+};
+export const CheckBox: ForwardRefExoticComponent<CheckBoxProps>;
