+++ packages/fuselage/src/components/Chip/index.d.ts
@@ -1,0 +1,8 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type ChipProps = ComponentProps<typeof Box>;
+const Chip: ForwardRefExoticComponent<ChipProps>;
+
+export = Chip;
