+++ packages/fuselage/src/components/Divider/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type DividerProps = ComponentProps<typeof Box>;
+export const Divider: ForwardRefExoticComponent<DividerProps>;
