+++ packages/fuselage/src/components/FieldGroup/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type FieldGroupProps = ComponentProps<typeof Box>;
+export const FieldGroup: ForwardRefExoticComponent<FieldGroupProps>;
