+++ packages/fuselage/src/components/Icon/index.d.ts
@@ -1,0 +1,8 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type IconProps = Omit<ComponentProps<typeof Box>, 'size'> & {
+  size?: ComponentProps<typeof Box>['width'];
+};
+export const Icon: ForwardRefExoticComponent<IconProps>;
