+++ packages/fuselage/src/components/InputBox/index.d.ts
@@ -1,0 +1,10 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type InputBoxProps = ComponentProps<typeof Box>;
+type InputBoxSkeletonProps = ComponentProps<typeof Box>;
+
+export const InputBox: ForwardRefExoticComponent<InputBoxProps> & {
+  Skeleton: ForwardRefExoticComponent<InputBoxSkeletonProps>;
+};
