+++ packages/fuselage/src/components/Modal/index.d.ts
@@ -1,0 +1,22 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type ModalProps = ComponentProps<typeof Box>;
+type ModalBackdropProps = ComponentProps<typeof Box>;
+type ModalCloseProps = ComponentProps<typeof Box>;
+type ModalContentProps = ComponentProps<typeof Box>;
+type ModalFooterProps = ComponentProps<typeof Box>;
+type ModalHeaderProps = ComponentProps<typeof Box>;
+type ModalThumbProps = ComponentProps<typeof Box>;
+type ModalTitleProps = ComponentProps<typeof Box>;
+
+export const Modal: ForwardRefExoticComponent<ModalProps> & {
+  Backdrop: ForwardRefExoticComponent<ModalBackdropProps>;
+  Close: ForwardRefExoticComponent<ModalCloseProps>;
+  Content: ForwardRefExoticComponent<ModalContentProps>;
+  Footer: ForwardRefExoticComponent<ModalFooterProps>;
+  Header: ForwardRefExoticComponent<ModalHeaderProps>;
+  Thumb: ForwardRefExoticComponent<ModalThumbProps>;
+  Title: ForwardRefExoticComponent<ModalTitleProps>;
+};
