+++ packages/fuselage/src/components/NumberInput/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type NumberInputProps = ComponentProps<typeof Box>;
+export const NumberInput: ForwardRefExoticComponent<NumberInputProps>;
