+++ packages/fuselage/src/components/PasswordInput/index.d.ts
@@ -1,0 +1,8 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type PasswordInputProps = ComponentProps<typeof Box> & {
+  error?: string;
+};
+export const PasswordInput: ForwardRefExoticComponent<PasswordInputProps>;
