+++ packages/fuselage/src/components/ProgressBar/index.d.ts
@@ -1,0 +1,9 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type ProgressBarProps = ComponentProps<typeof Box> & {
+  percentage?: number;
+  error?: string;
+};
+export const ProgressBar: ForwardRefExoticComponent<ProgressBarProps>;
