+++ packages/fuselage/src/components/Select/index.d.ts
@@ -1,0 +1,13 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type SelectOptions = readonly (readonly [string, string])[];
+
+type SelectProps = Omit<ComponentProps<typeof Box>, 'onChange'> & {
+  error?: string;
+  options: SelectOptions;
+  onChange: (value: SelectOptions[number][0]) => void;
+};
+
+export const Select: ForwardRefExoticComponent<SelectProps>;
