+++ packages/fuselage/src/components/Skeleton/index.d.ts
@@ -1,0 +1,8 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type SkeletonProps = ComponentProps<typeof Box> & {
+  variant?: 'rect';
+};
+export const Skeleton: ForwardRefExoticComponent<SkeletonProps>;
