+++ packages/fuselage/src/components/TextAreaInput/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type TextAreaInputProps = ComponentProps<typeof Box>;
+export const TextAreaInput: ForwardRefExoticComponent<TextAreaInputProps>;
