+++ packages/fuselage/src/components/TextInput/index.d.ts
@@ -1,0 +1,9 @@
+import { ComponentProps, ForwardRefExoticComponent, ReactNode } from 'react';
+
+import { Box } from '../Box';
+
+export type TextInputProps = ComponentProps<typeof Box> & {
+  addon?: ReactNode;
+  error?: string;
+};
+export const TextInput: ForwardRefExoticComponent<TextInputProps>;
