+++ packages/fuselage/src/components/Tile/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type TileProps = ComponentProps<typeof Box>;
+export const Tile: ForwardRefExoticComponent<TileProps>;
