+++ packages/fuselage/src/components/ToggleSwitch/index.d.ts
@@ -1,0 +1,6 @@
+import { ComponentProps, ForwardRefExoticComponent } from 'react';
+
+import { Box } from '../Box';
+
+type ToggleSwitchProps = ComponentProps<typeof Box>;
+export const ToggleSwitch: ForwardRefExoticComponent<ToggleSwitchProps>;
